refactor: EmitterRepository 패키지 경로 변경 및 관련 코드 수정
