+++ src/main/java/com/ootd/fitme/domain/notification/repository/NotificationInMemoryRepository.java
@@ -1,11 +1,9 @@
 package com.ootd.fitme.domain.notification.repository;
 
-import com.ootd.fitme.domain.notification.sse.EmitterRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
