+++ src/main/java/com/ootd/fitme/domain/notification/service/NotificationSseService.java
@@ -1,14 +1,13 @@
 package com.ootd.fitme.domain.notification.service;
 
 import com.ootd.fitme.domain.notification.dto.response.NotificationDto;
-import com.ootd.fitme.domain.notification.sse.EmitterRepository;
+import com.ootd.fitme.domain.notification.repository.EmitterRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
